Breaking this off because something is seriously wrong.

+++ src/glsl/atmosphere/atmosphere-pass.glsl
@@ -1,6 +1,6 @@
-precision mediump int;
-precision mediump float;
-precision mediump sampler3D;
+precision highp int;
+precision highp float;
+precision highp sampler3D;
 
 varying vec3 vWorldPosition;
 varying vec3 galacticCoordinates;
